feat: move conversation related functions to module
