+++ app/tests/modules/incident/test_incident.py
@@ -25,264 +25,6 @@ def test_is_floppy_disk_false():
     ), "The function should return False for reactions other than 'floppy_disk'"
 
 
-@patch("modules.incident.incident.open_modal")
-@patch("modules.incident.incident.log_to_sentinel")
-def test_handle_incident_action_buttons_call_incident(
-    _log_to_sentinel_mock, open_modal_mock
-):
-    client = MagicMock()
-    ack = MagicMock()
-    logger = MagicMock()
-    body = {
-        "actions": [
-            {
-                "name": "call-incident",
-                "value": "incident_id",
-                "type": "button",
-            }
-        ],
-        "user": {"id": "user_id"},
-    }
-    incident.handle_incident_action_buttons(client, ack, body, logger)
-    open_modal_mock.assert_called_with(client, ack, {"text": "incident_id"}, body)
-
-
-@patch("modules.incident.incident.webhooks.increment_acknowledged_count")
-@patch("modules.incident.incident.log_to_sentinel")
-def test_handle_incident_action_buttons_ignore(
-    _log_to_sentinel_mock, increment_acknowledged_count_mock
-):
-    client = MagicMock()
-    ack = MagicMock()
-    logger = MagicMock()
-    body = {
-        "actions": [
-            {
-                "name": "ignore-incident",
-                "value": "incident_id",
-                "type": "button",
-            }
-        ],
-        "channel": {"id": "channel_id"},
-        "user": {"id": "user_id"},
-        "original_message": {
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "foo",
-                    "text": "bar",
-                }
-            ],
-        },
-    }
-    incident.handle_incident_action_buttons(client, ack, body, logger)
-    increment_acknowledged_count_mock.assert_called_with("incident_id")
-    client.api_call.assert_called_with(
-        "chat.update",
-        json={
-            "channel": "channel_id",
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                    "text": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                }
-            ],
-        },
-    )
-
-
-@patch("modules.incident.incident.webhooks.increment_acknowledged_count")
-@patch("modules.incident.incident.log_to_sentinel")
-def test_handle_incident_action_buttons_ignore_drop_richtext_block(
-    _log_to_sentinel_mock,
-    increment_acknowledged_count_mock,
-):
-    client = MagicMock()
-    ack = MagicMock()
-    logger = MagicMock()
-    body = {
-        "actions": [
-            {
-                "name": "ignore-incident",
-                "value": "incident_id",
-                "type": "button",
-            }
-        ],
-        "channel": {"id": "channel_id"},
-        "user": {"id": "user_id"},
-        "original_message": {
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "foo",
-                    "text": "bar",
-                }
-            ],
-            "blocks": [
-                {
-                    "type": "rich_text",
-                    "block_id": "6Qv",
-                    "elements": [
-                        {
-                            "type": "rich_text_section",
-                            "elements": [{"type": "text", "text": "AWS notification"}],
-                        }
-                    ],
-                }
-            ],
-        },
-    }
-    incident.handle_incident_action_buttons(client, ack, body, logger)
-    increment_acknowledged_count_mock.assert_called_with("incident_id")
-    client.api_call.assert_called_with(
-        "chat.update",
-        json={
-            "channel": "channel_id",
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                    "text": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                }
-            ],
-            "blocks": [],
-        },
-    )
-
-
-@patch("modules.incident.incident.webhooks.increment_acknowledged_count")
-@patch("modules.incident.incident.log_to_sentinel")
-def test_handle_incident_action_buttons_ignore_drop_richtext_block_no_type(
-    _log_to_sentinel_mock,
-    increment_acknowledged_count_mock,
-):
-    client = MagicMock()
-    ack = MagicMock()
-    logger = MagicMock()
-    body = {
-        "actions": [
-            {
-                "name": "ignore-incident",
-                "value": "incident_id",
-                "type": "button",
-            }
-        ],
-        "channel": {"id": "channel_id"},
-        "user": {"id": "user_id"},
-        "original_message": {
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "foo",
-                    "text": "bar",
-                }
-            ],
-            "blocks": [
-                {
-                    "foo": "rich_text",
-                    "block_id": "6Qv",
-                    "elements": [
-                        {
-                            "type": "rich_text_section",
-                            "elements": [{"type": "text", "text": "AWS notification"}],
-                        }
-                    ],
-                }
-            ],
-        },
-    }
-    incident.handle_incident_action_buttons(client, ack, body, logger)
-    increment_acknowledged_count_mock.assert_called_with("incident_id")
-    client.api_call.assert_called_with(
-        "chat.update",
-        json={
-            "channel": "channel_id",
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                    "text": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                }
-            ],
-            "blocks": [
-                {
-                    "foo": "rich_text",
-                    "block_id": "6Qv",
-                    "elements": [
-                        {
-                            "type": "rich_text_section",
-                            "elements": [{"type": "text", "text": "AWS notification"}],
-                        }
-                    ],
-                }
-            ],
-        },
-    )
-
-
-# Test that the order of the ignore buttons are appended properly and the preview is moved up once the ignore button is clicked
-
-
-@patch("modules.incident.incident.webhooks.increment_acknowledged_count")
-@patch("modules.incident.incident.log_to_sentinel")
-def test_handle_incident_action_buttons_link_preview(
-    _log_to_sentinel_mock, increment_acknowledged_count_mock
-):
-    client = MagicMock()
-    ack = MagicMock()
-    logger = MagicMock()
-    body = {
-        "actions": [
-            {
-                "name": "ignore-incident",
-                "value": "incident_id",
-                "type": "button",
-            }
-        ],
-        "channel": {"id": "channel_id"},
-        "user": {"id": "user_id"},
-        "original_message": {
-            "attachments": [
-                {
-                    "color": "3AA3E3",
-                    "fallback": "foo",
-                    "text": "bar",
-                },
-                {
-                    "text": "test",
-                    "title": "title",
-                    "app_unfurl_url": "http://blah.com",
-                    "thumb_url": "http://blah.com/g/200/200",
-                    "image_url": "http://blah.com/g/200/200",
-                },
-            ],
-        },
-    }
-    incident.handle_incident_action_buttons(client, ack, body, logger)
-    increment_acknowledged_count_mock.assert_called_with("incident_id")
-    client.api_call.assert_called_with(
-        "chat.update",
-        json={
-            "channel": "channel_id",
-            "attachments": [
-                {
-                    "text": "test",
-                    "title": "title",
-                    "app_unfurl_url": "http://blah.com",
-                    "thumb_url": "http://blah.com/g/200/200",
-                    "image_url": "http://blah.com/g/200/200",
-                },
-                {
-                    "color": "3AA3E3",
-                    "fallback": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                    "text": "🙈  <@user_id> has acknowledged and ignored the incident.\n<@user_id> a pris connaissance et ignoré l'incident.",
-                },
-            ],
-        },
-    )
-
-
 @patch("modules.incident.incident.generate_incident_modal_view")
 @patch("modules.incident.incident.i18n")
 @patch("modules.incident.incident.slack_users.get_user_locale")
@@ -972,399 +714,6 @@ def test_incident_submit_does_not_invite_security_group_members_if_prefix_dev(
             call(channel="channel_id", users="creator_user_id"),
             call(channel="channel_id", users=["on_call_user_id"]),
         ]
-    )
-
-
-@patch("integrations.google_workspace.google_docs.extract_google_doc_id")
-def test_handle_reaction_added_floppy_disk_reaction_in_incident_channel(
-    mock_extract_google_doc_id,
-):
-    logger = MagicMock()
-    mock_client = MagicMock()
-
-    # Set up mock client and body to simulate the scenario
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "messages": [{"ts": "123456", "user": "U123456"}]
-    }
-    mock_client.users_profile_get.return_value = {"profile": {"real_name": "John Doe"}}
-    mock_client.bookmarks_list.return_value = {
-        "ok": True,
-        "bookmarks": [
-            {"title": "Incident report", "link": "https://docs.google.com/document/d/1"}
-        ],
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Assert the correct API calls were made
-    mock_client.conversations_info.assert_called_once()
-    mock_extract_google_doc_id.assert_called_once()
-
-
-def test_handle_reaction_added_non_incident_channel():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "general"}}
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Assert that certain actions are not performed for a non-incident channel
-    mock_client.conversations_history.assert_not_called()
-
-
-def test_handle_reaction_added_empty_message_list():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {"messages": []}
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Assert that the function tries to fetch replies when the message list is empty
-    mock_client.conversations_replies.assert_called_once()
-
-
-def test_handle_reaction_added_message_in_thread():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {"messages": []}
-    mock_client.conversations_replies.return_value = {
-        "messages": [{"ts": "123456", "user": "U123456"}]
-    }
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Assert that the function fetches thread replies
-    mock_client.conversations_replies.assert_called_once()
-
-
-def test_handle_reaction_added_incident_report_document_not_found():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    # Simulate no incident report document found
-    mock_client.bookmarks_list.return_value = {"ok": True, "bookmarks": []}
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    mock_client.users_profile_get.assert_not_called()
-
-
-def test_handle_reaction_added_adding_new_message_to_timeline():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "user": "U123ABC456",
-                "text": "Sample test message",
-                "ts": "1512085950.000216",
-            }
-        ],
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Make assertion that the function calls the correct functions
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-
-
-def test_handle_reaction_added_adding_new_message_to_timeline_user_handle():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "user": "U123ABC456",
-                "text": "<U123ABC456> says Sample test message",
-                "ts": "1512085950.000216",
-            }
-        ],
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Make assertion that the function calls the correct functions
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-
-
-def test_handle_reaction_added_returns_link():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "user": "U123ABC456",
-                "text": "<U123ABC456> says Sample test message",
-                "ts": "1512085950.000216",
-            }
-        ],
-    }
-    mock_client.chat_getPermalink.return_value = {
-        "ok": "true",
-        "channel": "C123456",
-        "permalink": "https://example.com",
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Make assertion that the function calls the correct functions
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-    mock_client.chat_getPermalink.assert_called_once()
-
-
-def test_handle_reaction_added_forwarded_message():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "attachments": [{"fallback": "This is a forwarded message"}],
-                "text": "Original message text",
-                "ts": "1617556890.000100",
-                "user": "U1234567890",
-                "files": [{"url_private": "https://example.com/image.png"}],
-            }
-        ],
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    incident.handle_reaction_added(mock_client, lambda: None, body, logger)
-
-    # Make assertion that the function calls the correct functions
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-
-
-def test_handle_reaction_removed_successful_message_removal():
-    # Mock the client and logger
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.users_profile_get.return_value = {
-        "profile": {"real_name": "John Doe", "display_name": "John"}
-    }
-    mock_client.bookmarks_list.return_value = {
-        "ok": True,
-        "bookmarks": [{"title": "Incident report", "link": "http://example.com"}],
-    }
-    mock_client.get_timeline_section.return_value = "Sample test message"
-    mock_client.replace_text_between_headings.return_value = True
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "user": "U123ABC456",
-                "text": "Sample test message",
-                "ts": "1512085950.000216",
-            }
-        ],
-    }
-
-    incident.handle_reaction_removed(mock_client, lambda: None, body, logger)
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-
-
-def test_handle_reaction_removed_successful_message_removal_user_id():
-    # Mock the client and logger
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.users_profile_get.return_value = {
-        "profile": {"real_name": "John Doe", "display_name": "John"}
-    }
-    mock_client.bookmarks_list.return_value = {
-        "ok": True,
-        "bookmarks": [{"title": "Incident report", "link": "http://example.com"}],
-    }
-    mock_client.get_timeline_section.return_value = "Sample test message"
-    mock_client.replace_text_between_headings.return_value = True
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-    mock_client.conversations_history.return_value = {
-        "ok": True,
-        "messages": [
-            {
-                "type": "message",
-                "user": "U123ABC456",
-                "text": "<U123ABC456> says Sample test message",
-                "ts": "1512085950.000216",
-            }
-        ],
-    }
-
-    incident.handle_reaction_removed(mock_client, lambda: None, body, logger)
-    mock_client.conversations_history.assert_called_once()
-    mock_client.bookmarks_list.assert_called_once()
-    mock_client.users_profile_get.assert_called_once()
-
-
-def test_handle_reaction_removed_message_not_in_timeline():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_info.return_value = {"channel": {"name": "incident-123"}}
-    mock_client.conversations_history.return_value = {
-        "messages": [{"ts": "123456", "user": "U123456"}]
-    }
-    mock_client.users_profile_get.return_value = {"profile": {"real_name": "John Doe"}}
-    mock_client.bookmarks_list.return_value = {
-        "ok": True,
-        "bookmarks": [{"title": "Incident report", "link": "http://example.com"}],
-    }
-    mock_client.get_timeline_section.return_value = "Some existing content"
-    mock_client.replace_text_between_headings.return_value = False
-
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-
-    assert (
-        incident.handle_reaction_removed(mock_client, lambda: None, body, logger)
-        is None
-    )
-
-
-def test_handle_reaction_removed_non_incident_channel_reaction_removal():
-    mock_client = MagicMock()
-
-    # Mock a non-incident channel
-    mock_client.conversations_info.return_value = {"channel": {"name": "general"}}
-
-    # Assert that the function does not proceed with reaction removal
-    mock_client.conversations_history.assert_not_called()
-
-
-def test_handle_reaction_removed_empty_message_list_handling():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_history.return_value = {"messages": []}
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-    assert (
-        incident.handle_reaction_removed(mock_client, lambda: None, body, logger)
-        is None
-    )
-
-
-def test_handle_reaction_removed_forwarded_message():
-    logger = MagicMock()
-    mock_client = MagicMock()
-    mock_client.conversations_history.return_value = {
-        "attachments": [{"fallback": "This is a forwarded message"}],
-        "text": "Original message text",
-        "ts": "1617556890.000100",
-        "user": "U1234567890",
-        "files": [{"url_private": "https://example.com/image.png"}],
-    }
-    body = {
-        "event": {
-            "reaction": "floppy_disk",
-            "item": {"channel": "C123456", "ts": "123456"},
-        }
-    }
-    assert (
-        incident.handle_reaction_removed(mock_client, lambda: None, body, logger)
-        is None
     )
